// src/app/Helpers/Enums/EnumDefinitions.ts
// ? For GENERAL
export enum ButtonTypes {
  submit = "submit",
  button = "button",
  reset = "reset",
}

// ? For REPORTS
export enum Genders {
  Male = "Erkek",
  Female = "Kadın",
  Other = "Diğer",
}

export enum BloodGroups {
  O = "O",
  A = "A",
  B = "B",
  AB = "AB",
}

export enum RhTypes {
  RhPositive = "Rh+",
  RhNegative = "Rh-",
}

export enum Reasons {
  assault = "Etkili Eylem",
  trafficAccident = "Trafik Kazası",
  workAccident = "İş Kazası",
  otherAccidents = "Diğer Kazalar",
  humanRights = "İnsan Hakları İhlali İddiası",
  suicide = "İntihar Girişimi",
  poision = "Zehirlenmeler",
}

export enum PeopleInRoom {
  doctor = "Tabip",
  staff = "Sağlık Personeli",
  other = "Sağlık Meslek Mensubu Personel",
  companion = "Refakatçi",
  security = "Güvenlik Görevlisi",
}